requirements.txt
django~=5.1.2
django-rest-framework
drf-spectacular~=0.27.2


djangorestframework~=3.15.2